Refactor ids

+++ resource/src/buffer.rs
@@ -3,14 +3,17 @@
 pub use gfx_hal::buffer::*;
 
 use {
-    crate::memory::{Block, Heaps, MappedRange, MemoryBlock, MemoryUsage},
+    crate::{
+        memory::{Block, Heaps, MappedRange, MemoryBlock, MemoryUsage},
+        util::{device_owned, Device, DeviceId},
+    },
     gfx_hal::{Backend, Device as _},
     relevant::Relevant,
 };
 
 /// Buffer info.
 #[derive(Clone, Copy, Debug)]
-pub struct Info {
+pub struct BufferInfo {
     /// Buffer size.
     pub size: u64,
 
@@ -25,11 +28,14 @@ pub struct Info {
 /// `B` - raw image type.
 #[derive(Debug)]
 pub struct Buffer<B: Backend> {
+    device: DeviceId,
     raw: B::Buffer,
     block: MemoryBlock<B>,
-    info: Info,
+    info: BufferInfo,
     relevant: Relevant,
 }
+
+device_owned!(Buffer<B>);
 
 impl<B> Buffer<B>
 where
@@ -37,9 +43,9 @@ where
 {
     /// Create buffer, allocate memory block for it and bind.
     pub unsafe fn create(
-        device: &B::Device,
+        device: &Device<B>,
         heaps: &mut Heaps<B>,
-        info: Info,
+        info: BufferInfo,
         memory_usage: impl MemoryUsage,
     ) -> Result<Self, failure::Error> {
         log::trace!("{:#?}@{:#?}", info, memory_usage);
@@ -57,6 +63,7 @@ where
         device.bind_buffer_memory(block.memory(), block.range().start, &mut buf)?;
 
         Ok(Buffer {
+            device: device.id(),
             raw: buf,
             block,
             info,
@@ -64,29 +71,37 @@ where
         })
     }
 
-    pub unsafe fn dispose(self, device: &B::Device, heaps: &mut Heaps<B>) {
+    /// Dispose of buffer resource.
+    /// Deallocate memory block.
+    pub unsafe fn dispose(self, device: &Device<B>, heaps: &mut Heaps<B>) {
+        self.assert_device_owner(device);
         device.destroy_buffer(self.raw);
         heaps.free(device, self.block);
         self.relevant.dispose();
     }
 
+    /// Get reference to raw buffer resource
     pub fn raw(&self) -> &B::Buffer {
         &self.raw
     }
 
+    /// Get mutable reference to raw buffer resource
     pub unsafe fn raw_mut(&mut self) -> &mut B::Buffer {
         &mut self.raw
     }
 
+    /// Get reference to memory block occupied by buffer.
     pub fn block(&self) -> &MemoryBlock<B> {
         &self.block
     }
 
+    /// Get mutable reference to memory block occupied by buffer.
     pub unsafe fn block_mut(&mut self) -> &mut MemoryBlock<B> {
         &mut self.block
     }
 
-    pub fn info(&self) -> &Info {
+    /// Get buffer info.
+    pub fn info(&self) -> &BufferInfo {
         &self.info
     }
 
@@ -104,7 +119,7 @@ where
     /// Map range of the buffer to the CPU accessible memory.
     pub fn map<'a>(
         &'a mut self,
-        device: &impl gfx_hal::Device<B>,
+        device: &Device<B>,
         range: std::ops::Range<u64>,
     ) -> Result<MappedRange<'a, B>, gfx_hal::mapping::Error> {
         self.block.map(device, range)
