+++ resource/src/sampler/mod.rs
@@ -3,18 +3,23 @@
 mod cache;
 
 use {
+    crate::util::{device_owned, Device, DeviceId},
     gfx_hal::{image::SamplerInfo, Backend, Device as _},
     relevant::Relevant,
 };
 
 pub use crate::sampler::cache::SamplerCache;
 
+/// Generic sampler resource wrapper.
 #[derive(Debug)]
 pub struct Sampler<B: Backend> {
+    device: DeviceId,
     raw: B::Sampler,
     info: SamplerInfo,
     relevant: Relevant,
 }
+
+device_owned!(Sampler<B>);
 
 impl<B> Sampler<B>
 where
@@ -22,26 +27,31 @@ where
 {
     /// Create new sampler.
     pub unsafe fn create(
-        device: &B::Device,
+        device: &Device<B>,
         info: SamplerInfo,
     ) -> Result<Self, gfx_hal::device::AllocationError> {
         let raw = device.create_sampler(info.clone())?;
         Ok(Sampler {
+            device: device.id(),
             raw,
             info,
             relevant: Relevant,
         })
     }
 
-    pub unsafe fn dispose(self, device: &B::Device) {
+    /// Destroy sampler resource.
+    pub unsafe fn dispose(self, device: &Device<B>) {
+        self.assert_device_owner(device);
         device.destroy_sampler(self.raw);
         self.relevant.dispose();
     }
 
+    /// Get reference to raw sampler resource.
     pub fn raw(&self) -> &B::Sampler {
         &self.raw
     }
 
+    /// Get mutable reference to raw sampler resource.
     pub unsafe fn raw_mut(&mut self) -> &mut B::Sampler {
         &mut self.raw
     }
